Make sizes to be of type size_t and correct function arguments that should be Byte (as the numerous casts to Byte in the function calls show).  Put macro arguments in paranthesis.
Taken-from: FreeBSD
Author: harti

+++ usr.bin/make/var_modify.c
@@ -37,7 +37,7 @@
  *
  * @(#)var.c	8.3 (Berkeley) 3/19/94
  * $FreeBSD: src/usr.bin/make/var.c,v 1.16.2.3 2002/02/27 14:18:57 cjc Exp $
- * $DragonFly: src/usr.bin/make/Attic/var_modify.c,v 1.7 2004/12/16 00:17:05 okumoto Exp $
+ * $DragonFly: src/usr.bin/make/Attic/var_modify.c,v 1.8 2004/12/16 22:05:23 okumoto Exp $
  */
 
 #include    <ctype.h>
@@ -400,7 +400,7 @@ VarSubstitute(const char *word, Boolean addSpace, Buffer buf, void *patternp)
 	     * buffer.
 	     */
 	    Boolean done;
-	    int origSize;
+	    size_t origSize;
 
 	    done = FALSE;
 	    origSize = Buf_Size(buf);
